+++ Stripe/STPAPIClient.m
@@ -71,14 +71,14 @@ static NSString * const APIEndpointFPXStatus = @"fpx/bank_statuses";
 @implementation Stripe
 
 static NSArray<PKPaymentNetwork> *_additionalEnabledApplePayNetworks;
+static NSString *_defaultPublishableKey;
 
 + (void)setDefaultPublishableKey:(NSString *)publishableKey {
-    [STPAPIClient validateKey:publishableKey];
-    [STPPaymentConfiguration sharedConfiguration].publishableKey = publishableKey;
+    _defaultPublishableKey = [publishableKey copy];
 }
 
 + (NSString *)defaultPublishableKey {
-    return [STPPaymentConfiguration sharedConfiguration].publishableKey;
+    return _defaultPublishableKey;
 }
 
 @end
@@ -89,7 +89,6 @@ static NSArray<PKPaymentNetwork> *_additionalEnabledApplePayNetworks;
 
 @property (nonatomic, strong, readwrite) NSMutableDictionary<NSString *,NSObject *> *sourcePollers;
 @property (nonatomic, strong, readwrite) dispatch_queue_t sourcePollersQueue;
-@property (nonatomic, strong, readwrite) NSString *apiKey;
 
 // See STPAPIClient+Private.h
 
@@ -117,32 +116,34 @@ static NSArray<PKPaymentNetwork> *_additionalEnabledApplePayNetworks;
 }
 
 - (instancetype)init {
-    return [self initWithConfiguration:[STPPaymentConfiguration sharedConfiguration]];
-}
-
-- (instancetype)initWithPublishableKey:(NSString *)publishableKey {
-    STPPaymentConfiguration *config = [[STPPaymentConfiguration alloc] init];
-    config.publishableKey = [publishableKey copy];
-    return [self initWithConfiguration:config];
-}
-
-- (instancetype)initWithConfiguration:(STPPaymentConfiguration *)configuration {
-    NSString *publishableKey = [configuration.publishableKey copy];
-    if (publishableKey) {
-        [self.class validateKey:publishableKey];
-    }
     self = [super init];
     if (self) {
-        _apiKey = publishableKey;
         _apiURL = [NSURL URLWithString:APIBaseURL];
-        _configuration = configuration;
-        _stripeAccount = configuration.stripeAccount;
+        _configuration = [STPPaymentConfiguration sharedConfiguration];
         _sourcePollers = [NSMutableDictionary dictionary];
         _sourcePollersQueue = dispatch_queue_create("com.stripe.sourcepollers", DISPATCH_QUEUE_SERIAL);
         _urlSession = [NSURLSession sessionWithConfiguration:[self.class sharedUrlSessionConfiguration]];
+        _publishableKey = [Stripe defaultPublishableKey];
     }
     return self;
 }
+
+- (instancetype)initWithPublishableKey:(NSString *)publishableKey {
+    STPAPIClient *apiClient = [self init];
+    apiClient.publishableKey = publishableKey;
+    return apiClient;
+}
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-implementations"
+- (instancetype)initWithConfiguration:(STPPaymentConfiguration *)configuration {
+    // For legacy reasons, we'll support this initializer and use the deprecated configuration.{publishableKey, stripeAccount} properties
+    STPAPIClient *apiClient = [self init];
+    apiClient.publishableKey = configuration.publishableKey;
+    apiClient.stripeAccount = configuration.stripeAccount;
+    return apiClient;
+}
+#pragma clang diagnostic pop
 
 + (NSURLSessionConfiguration *)sharedUrlSessionConfiguration {
     static NSURLSessionConfiguration  *STPSharedURLSessionConfiguration;
@@ -175,12 +176,7 @@ static NSArray<PKPaymentNetwork> *_additionalEnabledApplePayNetworks;
 
 - (void)setPublishableKey:(NSString *)publishableKey {
     [self.class validateKey:publishableKey];
-    self.configuration.publishableKey = [publishableKey copy];
-    self.apiKey = [publishableKey copy];
-}
-
-- (NSString *)publishableKey {
-    return self.configuration.publishableKey;
+    _publishableKey = [publishableKey copy];
 }
 
 - (void)createTokenWithParameters:(NSDictionary *)parameters
@@ -259,7 +255,7 @@ static NSArray<PKPaymentNetwork> *_additionalEnabledApplePayNetworks;
 }
 
 - (NSDictionary<NSString *, NSString *> *)authorizationHeaderUsingEphemeralKey:(STPEphemeralKey *)ephemeralKey {
-    NSString *authorizationBearer = self.apiKey ?: @"";
+    NSString *authorizationBearer = self.publishableKey ?: @"";
     if (ephemeralKey != nil) {
         authorizationBearer = ephemeralKey.secret;
     }