+++ Tests/Tests/STPPaymentMethodCardWalletMasterpassTest.m
@@ -20,7 +20,7 @@
 
 - (void)testDecodedObjectFromAPIResponseMapping {
     // We reuse the visa checkout JSON because it's identical to the masterpass version
-    NSDictionary *response = [STPTestUtils jsonNamed:STPTestJSONPaymentMethod][@"card"][@"wallet"][@"visa_checkout"];
+    NSDictionary *response = [STPTestUtils jsonNamed:STPTestJSONPaymentMethodCard][@"card"][@"wallet"][@"visa_checkout"];
     STPPaymentMethodCardWalletMasterpass *masterpass = [STPPaymentMethodCardWalletMasterpass decodedObjectFromAPIResponse:response];
     XCTAssertNotNil(masterpass);
     XCTAssertEqualObjects(masterpass.name, @"Jenny");